# swagger.json
{
    "swagger":"2.0",
    "info":{
        "description":"This is a test server",
        "version":"1.0.0",
        "title":"KindergartenChiu"
    },
...
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths":{
...
        "/child/add":{
            "post":{
                "tags":["Child"],
                "parameters":[
                    {
                        "name":"Authorization",
                        "type":"string",
                        "in":"header",
                        "description":"Jwt sign by user's apiKey"
                    },
                    {
                        "name":"parents_phone",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Parent's phone"
                    },
                    {
                        "name":"phone",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Child's phone"
                    },
                    {
                        "name":"password",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Child's password"
                    },
                    {
                        "name":"name",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Child's name"
                    },
                    {
                        "name":"address",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Child's address"
                    }
                ],
                "responses":{
                    "200":{
                        "description": "OK",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    },
                    "401":{
                        "decription": "Bad token",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    },
                    "500":{
                        "decription": "Server fail",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    }
                }
            }
        },
        "/child/addParent":{
            "post":{
                "tags":["Child"],
                "parameters":[
                    {
                        "name":"Authorization",
                        "type":"string",
                        "in":"header",
                        "description":"Jwt sign by user's apiKey"
                    },
                    {
                        "name":"phone",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Parent's phone"
                    },
                    {
                        "name":"password",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Parent's password"
                    },
                    {
                        "name":"name",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Parent's name"
                    },
                    {
                        "name":"address",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Parent's name"
                    }
                ],
                "responses":{
                    "200":{
                        "description": "OK",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    },
                    "401":{
                        "decription": "Bad token",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    },
                    "500":{
                        "decription": "Server fail",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    }
                }
            }
        },
        "/child/location":{
            "get":{
                "tags":["Child"],
                "parameters":[
                    {
                        "name":"Authorization",
                        "type":"string",
                        "in":"header",
                        "description":"Jwt sign by user's apiKey"
                    },
                    {
                        "name":"phone",
                        "type":"string",
                        "in":"Authorization",
                        "description":"User's phone"
                    },
                    {
                        "name":"child_uuid",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Child's uuid"
                    }
                ],
                "responses":{
                    "200":{
                        "description": "OK",
                        "schema":{
                            "type":"object",
                            "properties":{
                                "location":{
                                    "type":"object",
                                    "properties":{
                                        "lat":{
                                            "type":"string",
                                            "description":"latitude"
                                        },
                                        "lng":{
                                            "type":"string",
                                            "description":"longitude"
                                        }
                                    }
                                },
                                "remaining_time":{
                                    "type":"string",
                                    "description":"second"
                                }
                            }
                        }
                    },
                    "401":{
                        "decription": "Bad token",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    },
                    "500":{
                        "decription": "Server fail",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    }
                }
            }
        },
...
        "/child/arrive":{
            "post":{
                "tags":["Child"],
                "parameters":[
                    {
                        "name":"Authorization",
                        "type":"string",
                        "in":"header",
                        "description":"Jwt sign by teacher's apiKey"
                    },
                    {
                        "name":"phone",
                        "type":"string",
                        "in":"Authorization",
                        "description":"User's phone"
                    },
                    {
                        "name":"child_uuid",
                        "type":"string",
                        "in":"Authorization",
                        "description":"Child's uuid"
                    }
                ],
                "responses":{
                    "200":{
                        "description": "OK",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    },
                    "401":{
                        "decription": "Bad token",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    },
                    "500":{
                        "decription": "Server fail",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    }
                }
            }
        },
        "/child/allPickupChildren":{
            "get":{
                "tags":["Child"],
                "parameters":[
                    {
                        "name":"Authorization",
                        "type":"string",
                        "in":"header",
                        "description":"Jwt sign by user's apiKey"
                    },
                    {
                        "name":"phone",
                        "type":"string",
                        "in":"Authorization",
                        "description":"user's phone"
                    },
                    {
                        "name":"days",
                        "type":"array",
                        "items":{
                            "type":"boolean"
                        },
                        "in":"Authorization",
                        "description":"Array map '0~6' to 'Sunday ~ Saturday'"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type":"object",
                            "properties":{
                                "pickup":{
                                    "type":"array",
                                    "items":{
                                        "type":"object",
                                        "properties":{
                                            "uuid": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500":{
                        "decription": "Server fail",
                        "schema":{
                            "$ref": "#/definitions/message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions":{
...
        "message":{
            "type":"object",
            "properties":{
                "type":{
                    "type": "string"
                },
                "message":{
                    "type": "string"
                }
            }
        }
    }
}